+++ CHANGELOG.md
@@ -19,7 +19,7 @@ FlatLaf Change Log
     menu bar and main content. If enabled with `UIManager.put(
     "TitlePane.unifiedBackground", true );` then window title bar and menu bar
     use same background color as main content. (PR #268; issue #254)
-- JIDE Common Layer: Support `JideButton` and `JideToggleButton`.
+- JIDE Common Layer: Support `JideButton`, `JideLabel` and `JideToggleButton`.
 
 #### Fixed bugs